FlashVSR/FlashVSR/requirements.txt
torch
torchaudio
torchmetrics==1.7.3
torchsde==0.2.6
torchvision
accelerate==1.8.1
einops==0.8.1
huggingface-hub==0.34.4
matplotlib==3.10.3
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
peft==0.16.0
pillow==11.0.0
safetensors==0.5.3
sentencepiece==0.2.0
transformers==4.46.2
pytorch-lightning==2.5.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
protobuf==3.20.3
ftfy==6.3.1
pandas==2.3.0
tqdm
datasets